fix: select with header component can not scroll on android & use built-in overlay instead of react-native-portalize
