+++ packages/components/src/Select/Container/Mobile.tsx
@@ -1,12 +1,13 @@
-import React, { isValidElement } from 'react';
+import React, { isValidElement, useEffect, useMemo, useRef } from 'react';
 
-import Modal from 'react-native-modal';
+import { Animated } from 'react-native';
+import { Modalize } from 'react-native-modalize';
 
 import Box from '../../Box';
 import Button from '../../Button';
 import IconButton from '../../IconButton';
-import { useSafeAreaInsets } from '../../Provider/hooks';
-import ScrollView from '../../ScrollView';
+import { OverlayContainer } from '../../OverlayContainer';
+import { useSafeAreaInsets, useThemeValue } from '../../Provider/hooks';
 import Typography from '../../Typography';
 
 import { renderOptions } from './Option';
@@ -14,7 +15,7 @@ import { renderOptions } from './Option';
 import type { ChildProps } from '..';
 
 function Mobile<T>({
-  dropdownProps,
+  // dropdownProps,
   toggleVisible,
   options,
   onChange,
@@ -28,93 +29,135 @@ function Mobile<T>({
   renderItem,
   onModalHide,
   activatable,
+  withReactModal,
 }: ChildProps<T>) {
   const { bottom } = useSafeAreaInsets();
-  return (
-    <Modal
-      propagateSwipe
-      hideModalContentWhileAnimating
-      useNativeDriver
-      useNativeDriverForBackdrop
-      isVisible={!!visible}
-      onModalHide={onModalHide}
-      swipeDirection={['down']}
-      onSwipeComplete={toggleVisible}
-      onBackdropPress={toggleVisible}
-      animationInTiming={150}
-      animationOutTiming={150}
-      animationIn="slideInUp"
-      animationOut="slideOutDown"
-      style={{
-        justifyContent: 'flex-end',
-        margin: 0,
+  const modalizeRef = useRef<Modalize>(null);
+  const animated = useRef(new Animated.Value(0)).current;
+  const [defaultBgColor, handleBgColor] = useThemeValue([
+    'background-default',
+    'icon-subdued',
+  ]);
+
+  useEffect(() => {
+    if (visible) {
+      modalizeRef.current?.open();
+    } else {
+      onModalHide?.();
+      modalizeRef.current?.close();
+    }
+  }, [visible, onModalHide]);
+
+  const headerComponent = useMemo(
+    () => (
+      <Box
+        py="1"
+        px="2"
+        display="flex"
+        flexDirection="row"
+        justifyContent="space-between"
+        alignItems="center"
+        bg="transparent"
+      >
+        {/* placeholder */}
+        <Box width="12" />
+        <Box flex="1" flexDirection="row" justifyContent="center">
+          <Typography.Heading numberOfLines={1}>{title}</Typography.Heading>
+        </Box>
+        <IconButton
+          name="CloseOutline"
+          type="plain"
+          size="xl"
+          onPress={toggleVisible}
+          circle
+        />
+      </Box>
+    ),
+    [title, toggleVisible],
+  );
+
+  const footerComponent = useMemo(
+    () =>
+      isValidElement(footer) || footer === null ? (
+        footer
+      ) : (
+        <Box pb={`${bottom}px`} bg="surface-subdued">
+          <Box px="4" py="2" borderTopWidth={1} borderTopColor="border-subdued">
+            <Button
+              size="xl"
+              type="plain"
+              leftIconName={footerIcon}
+              onPress={onPressFooter}
+            >
+              {footerText}
+            </Button>
+          </Box>
+        </Box>
+      ),
+    [bottom, footer, footerIcon, footerText, onPressFooter],
+  );
+
+  const content = (
+    <Modalize
+      adjustToContentHeight
+      panGestureComponentEnabled
+      modalTopOffset={300}
+      withReactModal={withReactModal}
+      panGestureAnimatedValue={animated}
+      ref={modalizeRef}
+      withHandle={false}
+      onClose={onModalHide}
+      openAnimationConfig={{
+        timing: {
+          duration: 150,
+        },
+      }}
+      closeAnimationConfig={{
+        timing: {
+          duration: 150,
+        },
+      }}
+      modalStyle={{
+        backgroundColor: defaultBgColor,
+        borderTopLeftRadius: 24,
+        borderTopRightRadius: 24,
+      }}
+      handlePosition="inside"
+      handleStyle={{
+        alignSelf: 'center',
+        top: 8,
+        width: 45,
+        height: 5,
+        borderRadius: 5,
+        backgroundColor: handleBgColor,
+      }}
+      HeaderComponent={headerComponent}
+      FooterComponent={footerComponent}
+      childrenStyle={{
+        minWidth: '100%',
+        borderTopLeftRadius: 24,
+        borderTopRightRadius: 24,
+        paddingBottom: 0,
+      }}
+      scrollViewProps={{
+        contentContainerStyle: { padding: 2, paddingBottom: 4 },
       }}
     >
-      <Box
-        maxHeight="70%"
-        minHeight="180px"
-        minW="full"
-        bg="surface-subdued"
-        borderTopRadius="24px"
-        pb={`${footer === null ? bottom : 0}px`}
-        {...dropdownProps}
-      >
-        <Box
-          py="1"
-          px="2"
-          display="flex"
-          flexDirection="row"
-          justifyContent="space-between"
-          alignItems="center"
-          borderBottomColor="border-subdued"
-          borderBottomWidth={title ? 1 : undefined}
-        >
-          {/* placeholder */}
-          <Box width="12" />
-          <Box flex="1" flexDirection="row" justifyContent="center">
-            <Typography.Heading numberOfLines={1}>{title}</Typography.Heading>
-          </Box>
-          <IconButton
-            name="CloseOutline"
-            type="plain"
-            size="xl"
-            onPress={toggleVisible}
-            circle
-          />
-        </Box>
-        <ScrollView _contentContainerStyle={{ padding: 2, paddingBottom: '4' }}>
-          {renderOptions<T>({
-            options,
-            activeOption,
-            renderItem,
-            onChange,
-            activatable,
-          })}
-        </ScrollView>
-        {isValidElement(footer) || footer === null ? (
-          footer
-        ) : (
-          <Box pb={`${bottom}px`}>
-            <Box
-              px="4"
-              py="2"
-              borderTopWidth={1}
-              borderTopColor="border-subdued"
-            >
-              <Button
-                size="xl"
-                type="plain"
-                leftIconName={footerIcon}
-                onPress={onPressFooter}
-              >
-                {footerText}
-              </Button>
-            </Box>
-          </Box>
-        )}
-      </Box>
-    </Modal>
+      {renderOptions<T>({
+        options,
+        activeOption,
+        renderItem,
+        onChange,
+        activatable,
+      })}
+    </Modalize>
   );
+
+  if (withReactModal) {
+    return content;
+  }
+
+  return <OverlayContainer>{content}</OverlayContainer>;
 }
 
 export default Mobile;
