+++ packages/kit/src/provider/NavigationProvider.tsx
@@ -7,7 +7,6 @@ import {
   NavigationContainerRef,
 } from '@react-navigation/native';
 import { createURL } from 'expo-linking';
-import { Host } from 'react-native-portalize';
 
 import { Box, DialogManager, useThemeValue } from '@onekeyhq/components';
 import Toast from '@onekeyhq/components/src/Toast/Custom';
@@ -98,11 +97,9 @@ const NavigationApp = () => {
         theme={navigationTheme}
         linking={enableLinkingRoute ? linking : undefined}
       >
-        <Host>
-          <OverlayProvider>
-            <RootStack />
-          </OverlayProvider>
-        </Host>
+        <OverlayProvider>
+          <RootStack />
+        </OverlayProvider>
       </NavigationContainer>
       <Box
         overflow="hidden"
